Move service implementations into internal package (#957)
