+++ rest-api/src/main/java/gov/cms/qpp/conversion/api/services/internal/InOrderActionService.java
@@ -1,4 +1,4 @@
-package gov.cms.qpp.conversion.api.services;
+package gov.cms.qpp.conversion.api.services.internal;
 
 import java.util.concurrent.CompletableFuture;
