+++ rest-api/src/main/java/gov/cms/qpp/conversion/api/services/internal/ManifestVersionService.java
@@ -1,8 +1,10 @@
-package gov.cms.qpp.conversion.api.services;
+package gov.cms.qpp.conversion.api.services.internal;
 
 import org.springframework.stereotype.Service;
 
 import com.jcabi.manifests.Manifests;
+
+import gov.cms.qpp.conversion.api.services.VersionService;
 
 /**
  * Implementation of VersionService using META-INF/MANIFEST.mf
