+++ rest-api/src/main/java/gov/cms/qpp/conversion/api/services/internal/StorageServiceImpl.java
@@ -1,4 +1,4 @@
-package gov.cms.qpp.conversion.api.services;
+package gov.cms.qpp.conversion.api.services.internal;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.GetObjectRequest;
@@ -17,6 +17,7 @@ import org.springframework.stereotype.Service;
 
 import gov.cms.qpp.conversion.api.exceptions.UncheckedInterruptedException;
 import gov.cms.qpp.conversion.api.model.Constants;
+import gov.cms.qpp.conversion.api.services.StorageService;
 
 import java.io.InputStream;
 import java.util.Objects;
