+++ rest-api/src/test/java/gov/cms/qpp/conversion/api/services/internal/InOrderAsyncActionServiceTest.java
@@ -1,4 +1,4 @@
-package gov.cms.qpp.conversion.api.services;
+package gov.cms.qpp.conversion.api.services.internal;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -10,6 +10,7 @@ import org.springframework.core.task.TaskExecutor;
 import org.springframework.retry.backoff.FixedBackOffPolicy;
 import org.springframework.retry.support.RetryTemplate;
 
+import gov.cms.qpp.conversion.api.services.internal.InOrderActionService;
 import gov.cms.qpp.test.MockitoExtension;
 
 import java.util.concurrent.CompletableFuture;
